// types.ts
export interface Venue {
  id: string;
  name: string;
  location: string;
  vibe: string;
  features: string[];
  capacity: string;
  advantage: string;
  pricing: string;
  image: string;
  details: string;
}

export interface Service {
  title: string;
  description: string;
  icon: string;
}